refactor: tablegroup 과 table 을 같은 모듈에 위치해라
