--- module-order/src/main/java/kitchenpos/order/application/OrderTableStatusServiceImpl.java
@@ -1,28 +1,0 @@
-package kitchenpos.order.application;
-
-import java.util.Arrays;
-import kitchenpos.core.exception.CannotUpdateException;
-import kitchenpos.core.exception.ExceptionType;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.application.OrderTableStatusService;
-import kitchenpos.table.domain.OrderTable;
-import org.springframework.stereotype.Component;
-
-@Component
-public class OrderTableStatusServiceImpl implements OrderTableStatusService {
-
-    private final OrderRepository orderRepository;
-
-    public OrderTableStatusServiceImpl(OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-    }
-
-    @Override
-    public void validateOrderTableStatus(OrderTable orderTable) {
-        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
-            orderTable.getId(), Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new CannotUpdateException(ExceptionType.CAN_NOT_UPDATE_TABLE_IN_COOKING_AND_MEAL_STATUS);
-        }
-    }
-}
