+++ module-order/src/main/java/kitchenpos/order/event/TableEmptyChangeHandler.java
@@ -1,0 +1,36 @@
+package kitchenpos.order.event;
+
+import java.util.Arrays;
+import kitchenpos.core.exception.CannotUpdateException;
+import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.event.TableChangeEmptyEventPublisher;
+import org.springframework.context.event.EventListener;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+public class TableEmptyChangeHandler {
+
+    private final OrderRepository orderRepository;
+
+    public TableEmptyChangeHandler(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    @EventListener
+    @Transactional
+    public void handle(TableChangeEmptyEventPublisher eventPublisher) {
+        OrderTable orderTable = eventPublisher.getOrderTable();
+        validateOrderTableStatus(orderTable);
+    }
+
+    private void validateOrderTableStatus(OrderTable orderTable) {
+        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
+            orderTable.getId(), Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new CannotUpdateException(ExceptionType.CAN_NOT_UPDATE_TABLE_IN_COOKING_AND_MEAL_STATUS);
+        }
+    }
+}
