+++ module-order/src/main/java/kitchenpos/order/event/TableUnGroupEventHandler.java
@@ -1,0 +1,44 @@
+package kitchenpos.order.event;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.core.exception.CannotUpdateException;
+import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.event.TableUnGroupEventPublisher;
+import org.springframework.context.event.EventListener;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+public class TableUnGroupEventHandler {
+
+    private final OrderRepository orderRepository;
+
+    public TableUnGroupEventHandler(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    @EventListener
+    @Transactional
+    public void handle(TableUnGroupEventPublisher event) {
+        List<OrderTable> orderTables = event.getOrderTables();
+        validateOrderTableStatus(orderTables);
+        orderTables.forEach(OrderTable::unGroup);
+    }
+
+    private void validateOrderTableStatus(List<OrderTable> orderTables) {
+        List<Long> orderTableIds = orderTables.stream()
+            .map(OrderTable::getId)
+            .collect(Collectors.toList());
+
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+            orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new CannotUpdateException(
+                ExceptionType.CAN_NOT_UPDATE_TABLE_IN_COOKING_AND_MEAL_STATUS);
+        }
+    }
+}
