+++ module-table/build.gradle
@@ -5,3 +5,10 @@ bootJar {
 jar {
     enabled = true
 }
+
+dependencies {
+    testImplementation project(':module-order')
+
+    testImplementation 'org.springframework.boot:spring-boot-starter-web'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+}
