--- module-table/src/main/java/kitchenpos/table/application/OrderTableStatusService.java
@@ -1,8 +1,0 @@
-package kitchenpos.table.application;
-
-import kitchenpos.table.domain.OrderTable;
-
-public interface OrderTableStatusService {
-
-    void validateOrderTableStatus(OrderTable orderTable);
-}
