+++ module-table/src/test/java/kitchenpos/table/ModuleTableApplicationTests.java
@@ -1,0 +1,11 @@
+package kitchenpos.table;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModuleTableApplicationTests {
+    @Test
+    public void contextLoads() {
+    }
+}
