--- module-tablegroup/src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,7 +1,0 @@
-package kitchenpos.tablegroup.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
-
-}
