--- module-tablegroup/src/main/java/kitchenpos/tablegroup/domain/TableUnGroupEventPublisher.java
@@ -1,16 +1,0 @@
-package kitchenpos.tablegroup.domain;
-
-import java.util.List;
-import kitchenpos.table.domain.OrderTable;
-
-public class TableUnGroupEventPublisher {
-    private final List<OrderTable> orderTables;
-
-    public TableUnGroupEventPublisher(List<OrderTable> orderTables) {
-        this.orderTables = orderTables;
-    }
-
-    public List<OrderTable> getOrderTables() {
-        return orderTables;
-    }
-}
