--- module-tablegroup/src/main/java/kitchenpos/tablegroup/dto/response/TableGroupResponse.java
@@ -1,46 +1,0 @@
-package kitchenpos.tablegroup.dto.response;
-
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.dto.response.OrderTableResponse;
-import kitchenpos.tablegroup.domain.TableGroup;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class TableGroupResponse {
-
-    private Long id;
-    private LocalDateTime createdDate;
-    private List<OrderTableResponse> orderTables;
-
-    private TableGroupResponse(Long id, LocalDateTime createdDate,
-        List<OrderTableResponse> orderTables) {
-        this.id = id;
-        this.createdDate = createdDate;
-        this.orderTables = orderTables;
-    }
-
-    public TableGroupResponse() {
-    }
-
-    public static TableGroupResponse toResponse(TableGroup tableGroup, List<OrderTable> orderTables) {
-        List<OrderTableResponse> orderTableResponses = orderTables.stream()
-            .map(OrderTableResponse::of)
-            .collect(Collectors.toList());
-
-        return new TableGroupResponse(tableGroup.getId(), tableGroup.getCreatedDate(), orderTableResponses);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public List<OrderTableResponse> getOrderTables() {
-        return orderTables;
-    }
-}
