--- module-tablegroup/src/test/java/kitchenpos/tablegroup/ModuleTableGroupApplicationTests.java
@@ -1,11 +1,0 @@
-package kitchenpos.tablegroup;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ModuleTableGroupApplicationTests {
-    @Test
-    public void contextLoads() {
-    }
-}
