--- module-tablegroup/src/test/resources/application.properties
@@ -1,6 +1,0 @@
-logging.level.org.springframework.jdbc.core=TRACE
-spring.h2.console.enabled=true
-
-spring.jpa.hibernate.ddl-auto=create
-spring.jpa.show-sql=true
-logging.level.root=DEBUG
